+++ typescript/zetachain/zetacore/observer/operational_pb.d.ts
@@ -28,6 +28,15 @@ export declare class OperationalFlags extends Message<OperationalFlags> {
    */
   signerBlockTimeOffset?: Duration;
 
+  /**
+   * Minimum version of zetaclient that is allowed to run. This must be either
+   * a valid semver string (v23.0.1) or empty. If empty, all versions are
+   * allowed.
+   *
+   * @generated from field: string minimum_version = 3;
+   */
+  minimumVersion: string;
+
   constructor(data?: PartialMessage<OperationalFlags>);
 
   static readonly runtime: typeof proto3;